Update patch set 3
Patch Set 3:

> After running karaf =>
 > Caused by: java.lang.ClassNotFoundException: io.netty.channel.ChannelFutureListener
 > at org.eclipse.osgi.internal.loader.BundleLoader.findClassInternal(BundleLoader.java:501)
 > at org.eclipse.osgi.internal.loader.BundleLoader.findClass(BundleLoader.java:421)
 > at org.eclipse.osgi.internal.loader.BundleLoader.findClass(BundleLoader.java:412)
 > at org.eclipse.osgi.internal.baseadaptor.DefaultClassLoader.loadClass(DefaultClassLoader.java:107)

Good catch, thanks! I should have known that tweaking the bundle imports was unlikely to be the right solution. It turns out that the odl-netty feature ends up fighting with tcpmd5, so we need to migrate the latter to odl-netty before we can migrate sxp to odl-netty...

Patch-set: 3
Label: Verified=0