Ignore unit tests which rely on ordering of values in HashMap
Currently, the unit tests are failing,
preventing merge jobs from uploading new artifacts.
This is to prevent Nexus from forgetting the artifacts altogether,
which would break distribution jobs.

The correct fix would be for the tests to perform some king of parsing,
so that the final assert does not depend on order of bindings listed.

Change-Id: I998f048c564400740667273aeb27d9451ff92312
Signed-off-by: Vratko Polak <vrpolak@cisco.com>