partial fix
Partially fixed the md5 issue. Now the node start() method is waited upon in the nodeidentitiy listener, until a server is started. Then the method resumes and thus the md5 update does not fail on the uninitialized server. What remains to fix is the deadlock when multiple connections are added so the update method calls itself

Change-Id: If4f15353ae112146784009fc24389917a356a395
Signed-off-by: Martin Dindoffer <martin.dindoffer@pantheon.tech>