Delete entire node from Operational DS
This change alters the beaviour of deleting an SXP node from the
config DS. Previously, when a node (or an entire topo) was deleted
from the Config DS,  the change was applied to the operational DS
only partly. I.e. its non-sxp-augmented counterparts were left
untouched. This effectively created a "DB-memory leak" :
The SXP controller would write a new network topology node
stub to the DS when adding an SXP node, but would not
clean it up  when removing the node afterwards.

Change-Id: I6b8b0d4a315d521737ea529b285022bdf45ef596
Signed-off-by: Martin Dindoffer <martin.dindoffer@pantheon.tech>