Add a JTA spec dependency to satisfy Hazelcast in SFT
This change adds a JTA api spec as a compile time dependency to the sxp-core. For some reason, Karaf cannot resolve the javax.transaction.xa dependency of Hazelcast in the SFT, so we need to define it explicitly in the feature.xml.
This is a temporary (in reality permanent) workaround, until something somewhere gets fixed.

Change-Id: I9294cae60af27976f6e1c2d9157debe61ff0ba5e
JIRA: SXP-143
Signed-off-by: Martin Dindoffer <mdindoff@cisco.com>