Fix simultaneous connection problems and add ITs for Binding Origins
This commit fixes long standing bugs, where two nodes end up in
a corrupted state if they happen to connect to each other at the same time.
Fix is basically to just close the correct contexts and only when there is no
handshake in progress.
Also fixes the incorrect addition of the GLOBAL domain when there is one
provided during the construction of an SXP node.
Finally, it adds some IT for checking correct binding origin priorities
handling.

Change-Id: I29857bfc297c16cbe0b41135b7edbdc2c03641b5
Signed-off-by: Martin Dindoffer <mdindoff@cisco.com>
Signed-off-by: Martin Dindoffer <mdindoff@cisco.com>